+++ DollyCamPlugin2/interpstrategies/interpstrategy.h
@@ -10,7 +10,7 @@ protected:
 	void setCamPath(std::shared_ptr<savetype> path, int chaikinAmount);
 public:
 
-	virtual NewPOV GetPOV(float gameTime, int latestFrame) = 0;
+	virtual NewPOV GetPOV(float gameTime, float latestFrame) = 0;
 	virtual std::string GetName() = 0;
 };
 
@@ -22,7 +22,7 @@ class CosineInterpStrategy : public InterpStrategy
 {
 public:
 	CosineInterpStrategy(std::shared_ptr<savetype> _camPath);
-	virtual NewPOV GetPOV(float gameTime, int latestFrame);
+	virtual NewPOV GetPOV(float gameTime, float latestFrame);
 	virtual std::string GetName();
 };
 
@@ -30,7 +30,7 @@ class HermiteInterpStrategy : public InterpStrategy
 {
 public:
 	HermiteInterpStrategy(std::shared_ptr<savetype> _camPath);
-	virtual NewPOV GetPOV(float gameTime, int latestFrame);
+	virtual NewPOV GetPOV(float gameTime, float latestFrame);
 	virtual std::string GetName();
 };
