+++ DollyCamPlugin2/interpstrategies/nbezierinterp.h
@@ -29,7 +29,7 @@ private:
 	Mem<uint64_t, calc_factorial> factorial;
 public:
 	NBezierInterpStrategy(std::shared_ptr<savetype> _camPath, int degree);
-	virtual NewPOV GetPOV(float gameTime, int latestFrame);
+	virtual NewPOV GetPOV(float gameTime, float latestFrame);
 	virtual std::string GetName();
 };
